Revert "Sourceforge merge"
This reverts commit 4e1d3edd0e665a28de3e0f8860cc5bad0c0b284c.

Really should not have commited the 'download' version of sigproc to the master branch! This was a silly mistake.

+++ mmzap.f
@@ -52,7 +52,7 @@ c
 c     Special sequence 2: f = 50.0*i
 c
       k = 1
-      do ii=1,fnyq/50. - 1
+      do ii=1,int(fnyq/50.) - 1
          freq=ii*50.
          spseqf(2,k)=freq
          spseqw(2,k)=0.2
